[bug] Load playlist with error
When the on-disk paths for the song has changed, the existing behavior
was to fail loading the playlist. There was no feedback or error
reported to the user. This is extremely difficult to manage especially
when large collection of songs is reorganized.

The current patch continues to load the playlist even when songs are
missing. For those missing songs, playlist returned will contain the
orginal missing path with "-not-found.mp3" appended, the artist set to
"error artist" and album set to "error album".

Seeing this in the playlist, user can take corrective action, two of
which are
- restore the origian song at the path.
- remove the song from the playlist and save the playlist.